Rohan_Singhal/Assignment 2/Google-suite-main/Google-suite-main/requirements.txt
langchain
langchain-community
langchain-core
langgraph
langchain-groq
langchain-openai
google-auth 
google-auth-oauthlib
google-api-python-client
python-dotenv
requests